Add auth service
